Rename test package to be more explicit
